pyproject.toml
[build-system]
requires = ["setuptools>=77.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf-reporting"
version = "0.1.0"
description = "A package for generating PDF reports."
requires-python = ">=3.14"

dependencies = [
    "fpdf2>=2.8.5",
    "matplotlib>=3.7.0",
    "numpy==2.3.5",
    "matplotlib>=3.10.8"
]

[project.optional-dependencies]
dev = [
    "ruff>=0.14.10",
    "mypy>=1.8.0",
    "pytest>=8.0",
    "pytest-cov>=5.0"
]

[project.scripts]
lint = "fpdf_reporting.cli:lint"
format = "fpdf_reporting.cli:format"
typecheck = "fpdf_reporting.cli:typecheck"

[tool.ruff]
line-length = 88
target-version = "py314"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # allow re-exports
"tests/**" = ["S101"]  # allow asserts

[tool.mypy]
python_version = "3.14"
strict = true
mypy_path = ["src"]
explicit_package_bases = true
exclude = "tests/"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "--cov=fpdf_reporting --cov-report=term-missing"
testpaths = ["tests"]

[tool.coverage.run]
omit = [
    "*/__init__.py",
    "fpdf_reporting/cli.py",
]
branch = true
source = ["fpdf_reporting"]

[tool.coverage.report]
show_missing = true
fail_under = 80